Fixed the tests and moved them to separate files
