+++ packages/ogma-annotations/test/unit/utils.ts
@@ -1,0 +1,3 @@
+import Ogma from "@linkurious/ogma";
+
+export const createOgma = () => new Ogma({ renderer: "svg" });
